backend/requirements.txt
sqlalchemy==1.4.23
psycopg2-binary>=2.9.0
alembic>=1.7.0
python-dotenv==0.19.0
fastapi==0.68.1
uvicorn==0.15.0
# Use pydantic-settings v1 for compatibility with Pydantic v1 and FastAPI 0.68.1
pydantic-settings~=1.0.0
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
ultralytics>=8.0.0
opencv-python==4.5.3.56
numpy==1.21.2
torch>=2.0.0
torchvision>=0.15.0
face-recognition==1.3.0
pillow==8.3.2
python-magic==0.4.24
shutil==0.0.1
boto3>=1.20.0 # Added AWS SDK
deepface>=0.0.79 # Added for demographics analysis